Search & Files
Finally got to make it all work as it is supposed to, files work but there's a small problem I still have to work on related to downloading them. Authentication works smoothly and so on. I think I forgot to check on other routes besides /busqueda.

From now on we can focus on adding new stuff and making it look better on phones and every device.